Added .has and .values methods
Added .has and .values methods. These are from Set<Emitter> interface. 
EventMultiplexer actually implements both EventEmitter and Set interfaces for a reason: 
It's a repository of objects, which also handles their events in a consistent way.

Also, `values` method must be defined, because there should be a way to deal with objects
in it, references to which have been lost by user. At least it allows to clear the Multiplexer.